+++ packages/ragbits-evaluate/src/ragbits/evaluate/factories/__init__.py
@@ -39,6 +39,6 @@ async def _add_example_documents(document_search: DocumentSearch) -> None:
 
 def basic_document_search_factory() -> DocumentSearch:
     """A factory for basic example document search instance"""
-    document_search = DocumentSearch(embedder=LiteLLMEmbedder(), vector_store=InMemoryVectorStore())
+    document_search = DocumentSearch(vector_store=InMemoryVectorStore(embedder=LiteLLMEmbedder()))
     asyncio.run(_add_example_documents(document_search))
     return document_search