+++ packages/ragbits-evaluate/CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## Unreleased
+- Compability with the new Vector Store interface from ragbits-core (#288)
 
 ## 0.9.0 (2025-02-25)